+++ ft_puthex.c
@@ -6,13 +6,13 @@
 /*   By: alejogogi <alejogogi@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 13:04:48 by alejogogi         #+#    #+#             */
-/*   Updated: 2025/01/09 18:50:54 by alejogogi        ###   ########.fr       */
+/*   Updated: 2025/01/09 18:53:18 by alejogogi        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthex(int num, char c)
+int	ft_puthex(unsigned long long num, char c)
 {
 	char	*hex;
 	int		ln;
